feat(Sync): Include access token and consumer key in GraphQL requests
Due to quirks in the legacy system, the access token and consumer key
need to be included as query parameters for each request.

We do this by creating a custom `ApolloInterceptor` that modifies the
request to include the new values. The custom interceptor is then
provided by a simple `PrependingInterceptorProvider`.

Also: update GraphQL schema, list query, and auto-generated code.