Pull Sentry down into Sync
This is necessary because we'll need to do some Sentry-related work
inside of Sync, but because of the nature of Swift Package Manager, we
can only have one target that depends on a package. Thus, we have Sync
depend on the Sentry SDK and create a wrapper (Crashlogger) that can be
used within the app target.

This is a workaround until we resolve the dependency strategy issues
more broadly.