project: target iOS 15 and macOS 12
- Update swit package versions
- Also update usage of Apollo API to adapt to name change
- Fix navigation to article view
- Fix UI test XCUIElement query for web reader button

Temporarily use fix/xcode-13 branch of Kingfisher

There is a bug in current beta versions of Xcode that are causing this
error:

```
Failed to build module 'Accelerate'; this SDK is not supported by the compiler
```

It is documented here[fb] and there is an issue in the Kingfisher repo
tracking the problem. The workaround is to set the deployment target in
the swift package (which is being done in the fix/xcode-13 branch)

Once Xcode gets its act together, we can revert this change.

[fb]: https://developer.apple.com/forums/thread/682285
[kf-issue]: https://github.com/onevcat/Kingfisher/issues/1725