Recommendation cells: Increase tappable area of save button
We increase the tappable area of the button by setting the button's
contentInsets to something more reasonable (8pts on each edge) instead
of using a content inset of `.zero`.

In order to keep things aligned and spaced properly, we pin the edges of
the button to the edge of the cell (as opposed to the cell's layout
margin) and decrease the spacing between the excerpt text and the
button. These changes offset the additional space that is now occupied
by the button itself while maintaining the same visual alignment and
spacing.

https://getpocket.atlassian.net/browse/IN-252