Clarify launch args and clearing of various local storage
When the app is first launched, we store a value in user defaults so
that subsequent launches will know that the app has already been run at
least once.

When signing out, we want to clear all the user settings _except for
this one value_. Thus, we store the first launch flag in its own
UserDefaults with its own suite name. That way, we can easily clear all
other user settings without stepping on the first launch flag.

Also, when running unit tests, we need to be able to launch the app and
have it be in a predictable state. We do this by having a launch
argument for each type of local storage that we may want to clear out.
Having all these arguments can be confusing, so we encapsulate the
common scenarios into some pre-built instances of LaunchArguments (e.g.
"firstLaunch" and "bypassSignIn").

This allows us to think in higher level concepts as opposed to having to
think about which particular storage options we want to clear or
preserve.