Reader: Better image resizing/scaling
This one is a bit tricky, but the actual most meaningful change is the
omission of the `scaleFactor` option when setting the image.

Kingfisher recommends _against_ using the device's scale factor here,
`scaleFactor` is actually intended to be used when you know the image's
scale factor ahead of time.

Since we don't know the scale factor, we have to assume 1x. The end
result is that Kingfisher is much better at resizing images and they
look much better with this strategy.

Also included here:
- Fixup the identifier for the `OnlyResizeDownProcessor`
- `OnlyResizeDownProcessor` will only resize if the image is too wide to
  fit in the target size, since we don't need to limit vertical space.