Tweaks to image cell styling
We want to fill in empty space using a very light shade of grey when the
image is not wide enough to fill the view, but we do not want this
filler rectangle to have rounded corners.

Thus we round tell Kingfisher to round the corners on the image itself,
and remove the corner radius for the UIImageView.

Also, we do not want to show the background filler if the image _is_
wide enough (otherwise there would be tiny triangles in the corners). So
we add a bit of logic to toggle the background color on and off as
necessary.