Respond to changes in reader settings
Because presenters want to cache their content, they were not properly
responding to changes in reader settings.

Thus the presenters need to have a mechanism for clearing cached values
and we invoke this mechanism when reader settings and/or size classes
change.

That way, when the collection view is reloaded, the size and content
will be recalculated as necessary.