Sync tests: Only call source methods from main thread
PocketSource should not be considered thread safe! Whenever calling
methods on Source, we need to be sure to do so from the main thread.
This is because Source methods often work with CoreData and all of our
CoreData work needs to happen on the main thread.

This change removes the `async` keyword from a couple of places where it
was unnecessary and adds `@MainActor` to test methods that need to run
in an async context but on the main thread.