Persist and restore offline operations
If a user performs any actions while offline, and then the app is
terminiated, we need to be able to restore those operations on the next
launch of the application.

Operations are stored in core data via `SyncTask`. When `PocketSource`
is initialized, it will check to see if any items have been persisted
locally and re-enqueue those operations.

https://getpocket.atlassian.net/browse/IN-393