Operations: only do core data work on main thread
Since things can happen on background threads when using async/await, we
need to make sure core data work is happening on the main thread within
our operations.
We accomplish this by adding the `@MainActor` attribute to functions
that work with CoreData.

Also: Move `FetchArchivePageOperation` into the Operations directory so
it can live with its family.