Use AppSession to coordinate app state transitions
AppSession.$currentSession is either set as a valid value, or nil. RootCoordinator handles the setting of this session to determine which state the app should be in - logged out, or logged in. SessionController has been renamed to SessionListener and repurposed to clear out additional services when the currentSession is set to nil.