Remove schema.json
This file was originally put here to support code generation but we have
changed our strategy for generating code so we don't need to keep this
in source control.

Previously, we had a build step that would automatically run the Apollo
code generation workflow but it turned out to be problematic for a
number of reasons. Plus, we don't really need to run the codegen very
often. Now we just run the codegen process manually when needed.

What this means is that if the schema file isn't present on your
machine, you'll want to run the `ApolloCodegen` target with the
`download-schema` launch argument.