Remove unused ApolloCodegen target
We had created this a few months ago so that we could automatically
generate apollo queries during every build. We ended up removing this as
a dependency of Pocket because it was constantly failing for various
reasons.

Since we don't actually need to run codegen very often, we just manually
run ApolloCodegen package manually with the `download-schema` and/or
`generate` arguments.