Move PocketSaveService into Sync
This allows us to keep `Space` internal to the Sync framework. It also
eliminates a number of duplicated test support classes (mocks) that were
copied over from Sync.

Instead of creating the Space and Apollo client within the
MainViewController, we have a convience initializer on PocketSaveService
that can create those dependencies without exposing them to higher level
code. This convenience intializer accepts the SessionProvider and
consumer key that are necessary to initialize an apollo client.

This pattern is follows what we've done with PocketSource and PocketKit.