Update signing certificate
Our development certificate has expired, so I've generated a new one.
The last one was created by "Mozilla Release Engineering" user, back
before we had permission to create one for ourselves.

I created the new one myself, and Apple sets the name of the cert to the
name of the user creating it. Hence the different name.