Adjust TopicChipCell sizing and spacing
When calculating the size of a cell, we were using different hard-coded
values for margins/spacing, causing the buttons to be rendered poorly.

Also, we weren't accounting for whether the button has an image when
calculating the width.

1. Extract out some constants so we can set margins and calculate height
   using the same values.
2. Adjust sizing logic to account for presence of an image.

Also: The vertical margins around the topic buttons was a little
excessive, so I adjusted that slightly. (same goes for the filter
buttons on My List tab).

https://getpocket.atlassian.net/browse/IN-566