Textile: add per-font Markdown styling
Based on Figma designs, when Graphik is used, the emphasized weight should be medium, and not semibold. The style definitions for when Graphik are used have been updated to use medium, as well as `TextileStyler`, which will now set the appropriate weight when using Graphik is used as an updated style.