Decouple Space from PersistentContainer
Space objects are essentially wrappers around the a
NSManagedObjectContext but initializer for `Space` requires a
`PersistentContainer`, basically just that it can access the
`viewContext` property on the container.

Instead, we pass the instance of NSManagedObjectContext into Space.init.
This will allow us to have Space objects that are not bound to the main
thread (and can have child NSManagedObjectContexts).