Fix flaky test
This test fails intermittently because there are a few possible requests
that could be received while we're waiting on the request to delete an
item.
By adding handlers for each type of request– and only fulfilling our
expectation when we receive the exact request we're expecting, we can be
sure that our assertions are being made correctly.