Use consistent action identifier when attaching button actions
When adding a button action *that should replace another button action*,
it's crucial that the two actions have the same identifier.

Per Apple's documentation:
> UIActions are uniqued based on their identifier, and subsequent
> actions with the same identifier replace previously added actions.

To make sure that we don't attach more than one action to a button, we
use the `.recommendationPrimary` action identifier when attaching the
save/archive actions in home and slate details.

To keep things consistent, the following methods were renamed:

- HomeViewModel.[save/archive]Action(for:) -> primaryAction(for:)
- SlateDetailsViewModel.[save/archive]Action(for:) -> primaryAction(for:)

https://getpocket.atlassian.net/browse/IN-724