Improve stability and reliability of Home and other code cleanup (#216)
This change is tantamount to a rewrite of HomeViewModel. The impetus for the change was to eliminate the various usages of multiple NSFetchedResultsController objects to observe changes to core data models. Doing all of this observation within individual objects causes unexpected behavior due to an unpredictable order in which the change handlers are invoked.
Instead of having multiple handlers, the HomeViewModel now simply responds to the NSManagedObjectContext.didSaveObjectsNotification and determines any cells that need to be updated (e.g. a SavedItem that has been favorited or a Recommendation that has been saved).

See notes below for other changes to adjacent code.https://getpocket.atlassian.net/browse/IN-713
https://getpocket.atlassian.net/browse/IN-724
https://getpocket.atlassian.net/browse/IN-728