Coordinators: properly unwind presented/pushed viewmodels
The goal of our coordinators is twofold:

1. We want to be able to encapsulate all business logic within our
   view models, and simply observe their state to know when to present
   modal views or push views onto a navigation controller.
2. When users transition from compact to regular size classes (or vice
   versa), we can keep our compact view hiearchy in sync with the split
   view's hiearchy.

The theoretical concept is mostly good —containing state within view
models that are shared between both the regular and compact coordinators
allows us to easily reconstruct the view hiearchy when the user
transitions size classes.

However, the practical reality is...less good. One of the complications
with this approach is that we need to clean up after ourselves when
users dismiss modals or pop the navigation stack. For example, when the
share sheet is dismissed, we need to clear the `sharedActivity` property
of the view model that presented it.

To make this "unwinding" easier to keep track of, a set of methods to
clear certain properties have been added to objects throughout the view
model tree. This means the coordinators can call a single method on
their top level view model to "clearPresentedSharedActivity" and the
message will delegate down the tree to ensure that the correct property
is cleared, regardless of what level the shared activity was shared
from.

https://getpocket.atlassian.net/browse/IN-701