HomeCarouselItemCell: call super in `layoutSubviews` override
It's generally best practice to call the superclass's implementation of
a method that you're overriding _at some point_ within the override. In
this case, calling `super.layoutSubviews()` is necessary so that
constraints are re-evaluated and the layout gets updated when necessary
(e.g. when user rotates their device to a new orientation).