Avoid system alert when pasting URLs
Reading directly from the pasteboard triggers a system-wide alert,
asking the user to grant the app permission to read the contents of the
pasteboard. To avoid this, we use adopt usage of `UIPasteControl`, which
acts as a button that allows us to read from the clipboard without
requiring the user to grant permission.