refactor(analytics): request specific platform context properties
Be more aggressive about what we want Snowplow to automatically add as properties to mobile contexts. For example, we do not require knowing device storage, yet there was potential for it causing hangs within the Snowplow SDK: https://github.com/snowplow/snowplow-ios-tracker/issues/798. This list is based on what could be "skipped" when brought up to EM.