In 1597 Sharing PDF links to Pocket (#832)
* fix(extension) Continue looping through items when saving from extension

* fix(extension) Refactor code

* fix(extension) Fix warning

* fix(extension) Add happy-path test

* project(spm): update snowplow

* refactor(analytics): request specific platform context properties

Be more aggressive about what we want Snowplow to automatically add as properties to mobile contexts. For example, we do not require knowing device storage, yet there was potential for it causing hangs within the Snowplow SDK: https://github.com/snowplow/snowplow-ios-tracker/issues/798. This list is based on what could be "skipped" when brought up to EM.

* fix(extension) Fix linting issue

* fix(extension) Change to only save the first item encountered

---------

Co-authored-by: David Skuza <dskuza@mozilla.com>